Create #C - Examination project
This code was created as part of a programming course in C#. It is a console application that simulates a simple blogging platform. The program allows users to perform various actions, including viewing all posts, creating new posts with titles and content, searching for posts by title, clearing all posts, and exiting the program. The code utilizes object-oriented programming concepts with a class 'Me' to store user information and a list to store blog posts. It also includes error handling to guide users in case of incorrect menu selections. Feel free to explore and contribute to this C# project on GitHub!